Add `ModFlagNot` tag functionality
I had an issue when processing the improved "PerfectAgony" processing.
Generic non-ailment dots do not have their own `ModFlag` so there was no
way to mark a mod as only applying to "non-ailments".

The implementation of `ModFlagNot`, which is similar to `ModFlagOr` was
pretty straightforward and might be useful in the future, so I just
added it.

I also updated the respective docs