Make "Perfect Agony" actually convert `CritMulti`
The old processing added a single "Override" mod for "DotMultiplier"
that used a `Sum()` of `CritMultiplier` as the value. It mostly worked,
but meant that the breakdown for `Eff. Dot Multi` was pretty useless /
inaccurate.

This solution actually adds new `DotMultiplier` mods that are based on
`CritMultiplier` mods, but are exclusive to ailments. Old ailment mods
are deactivated, and non-ailment `DotMultiplier` mods are made exclusive
to non-ailments (duh)